fix-actions - Changed app.js to index.js and vice versa [Finishes#]
