+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon --exec babel-node src/app.js",
-    "start": "node dist/app.js",
+    "dev": "nodemon --exec babel-node src",
+    "start": "node dist/index.js",
     "build": "babel src -d dist",
     "lint": "esw src --color",
     "lint:fix": "npm run lint --fix",
