+++ src/app.js
@@ -1,13 +1,22 @@
-import dotenv from "dotenv";
-import app from "./index";
+import express from "express";
+import cors from "cors";
+import routes from "./routes/index";
 
+import languages from './locales/languages';
+import swaggerDocs from './api-docs/swagger';
 import connectDb from "./database/connectDb";
 
-dotenv.config();
+const app = express();
 
-const port = process.env.PORT ? process.env.PORT : 3000;
+app.use(express.json());
+app.use(cors({ origin: "*" }));
 
-app.listen(port, async () => {
-  await connectDb();
-  console.log(`Server is running on http://localhost:${port}`);
+swaggerDocs(app);
+languages(app);
+
+app.get("/", (_, res) => {
+  res.status(200).json("Welcome to our Ecommerce App");
 });
+app.use("/", routes);
+
+export default app;
